feat: refactor on the whole ia generating feature
